+++ src/Rocketeer/Tasks/Setup.php
@@ -76,17 +76,22 @@ class Setup extends Task
 	protected function createStages()
 	{
 		// Get stages
-		$stages = $this->rocketeer->getStages();
-		if (empty($stages)) {
-			$stages = array(null);
+		$availableStages = $this->rocketeer->getStages();
+		$originalStage   = $this->rocketeer->getStage();
+		if (empty($availableStages)) {
+			$availableStages = array(null);
 		}
 
 		// Create folders
-		foreach ($stages as $stage) {
+		foreach ($availableStages as $stage) {
 			$this->rocketeer->setStage($stage);
 			$this->createFolder('releases', true);
 			$this->createFolder('current', true);
 			$this->createFolder('shared', true);
+		}
+
+		if ($originalStage) {
+			$this->rocketeer->setStage($originalStage);
 		}
 	}
 }